move SelectionProvider tests over to Vitest
